Trang tri map

+++ Assets/Scripts/CollectableRotate.cs
@@ -2,7 +2,8 @@ using UnityEngine;
 
 public class CollectableRotate : MonoBehaviour
 {
-    [SerializeField] int rotateSpeed = 1;
+    [SerializeField] float rotateSpeed = 1;
+
 
     void Update()
     {
